permisos no admisn ocultacion de icono eliminar y modificar servicios

+++ resources/views/servicio/servicios.blade.php
@@ -48,12 +48,18 @@
                 <td class="text-center"><a href="{{route('servicio.show', $servicio->id)}}">{{config('selects.tipoServicio')[$servicio->tipo_servicio_id]}}</a></td>
                 <td class="text-center">{{$servicio->direccion}}</td>
                 <td class="text-center">{{\Carbon\Carbon::parse($servicio->hora_regreso)->format('d/m/Y H:i:s')}}</td>
-                <td class="text-center">
-                  {{ Form::open(['route' => ['servicio.destroy', $servicio->id], 'method' => 'delete']) }}
-                      <button type="submit" class="glyphicon glyphicon-trash"></button>
-                  {{ Form::close() }}
-                </td>
-                <td class="text-center"><a class="glyphicon glyphicon-edit" href="{{ route('servicio.edit', $servicio->id) }}"></a></td>
+                @if (Auth::user()->admin)
+                  <td class="text-center">
+                    {{ Form::open(['route' => ['servicio.destroy', $servicio->id], 'method' => 'delete']) }}
+                        <button type="submit" class="glyphicon glyphicon-trash"></button>
+                    {{ Form::close() }}
+                  </td>
+                  <td class="text-center"><a class="glyphicon glyphicon-edit" href="{{ route('servicio.edit', $servicio->id) }}"></a></td>
+                @else
+                  <td class="text-center" colspan="2">
+                    <button type="submit" class="glyphicon glyphicon-ban-circle" title="Sin permisos para eliminar/modificar"></button>
+                  </td>
+                @endif
               </tr>
             @endif
           @endforeach
